utilities, css-selector, test/css-selector: organize code a bit
