+++ test/css-selector.lua
@@ -1,60 +1,5 @@
 require "../css-selector"
 
-function deepcompare(t1,t2,ignore_mt)
-    local ty1 = type(t1)
-    local ty2 = type(t2)
-    if ty1 ~= ty2 then return false end
-    -- non-table types can be directly compared
-    if ty1 ~= 'table' and ty2 ~= 'table' then return t1 == t2 end
-    -- as well as tables which have the metamethod __eq
-    local mt = getmetatable(t1)
-    if not ignore_mt and mt and mt.__eq then return t1 == t2 end
-    for k1,v1 in pairs(t1) do
-        local v2 = t2[k1]
-        if v2 == nil or not deepcompare(v1,v2) then return false end
-    end
-    for k2,v2 in pairs(t2) do
-        local v1 = t1[k2]
-        if v1 == nil or not deepcompare(v1,v2) then return false end
-    end
-    return true
-end
-
-function table.val_to_str ( v )
-    if "string" == type( v ) then
-        v = string.gsub( v, "\n", "\\n" )
-        if string.match( string.gsub(v,"[^'\"]",""), '^"+$' ) then
-            return "'" .. v .. "'"
-        end
-        return '"' .. string.gsub(v,'"', '\\"' ) .. '"'
-    else
-        return "table" == type( v ) and table.tostring( v ) or tostring( v )
-    end
-end
-
-function table.key_to_str ( k )
-    if "string" == type( k ) and string.match( k, "^[_%a][_%a%d]*$" ) then
-        return k
-    else
-        return "[" .. table.val_to_str( k ) .. "]"
-    end
-end
-
-function table.tostring( tbl )
-    local result, done = {}, {}
-    for k, v in ipairs( tbl ) do
-        table.insert( result, table.val_to_str( v ) )
-        done[ k ] = true
-    end
-    for k, v in pairs( tbl ) do
-        if not done[ k ] then
-            table.insert( result,
-            table.key_to_str( k ) .. "=" .. table.val_to_str( v ) )
-        end
-    end
-    return "{" .. table.concat( result, "," ) .. "}"
-end
-
 --- Test triples
 --- ------------
